requirements.txt
torch>=2.2
torchvision>=0.17
open_clip_torch>=2.24.0
transformers>=4.46
PyYAML>=6.0
numpy>=1.24
Pillow>=10.0
tqdm>=4.66
httpx[socks]>=0.28.1
pytesseract>=0.3.10
openai>=1.51.0
